requirements.txt
Django>=4.1,<4.2
psycopg2-binary>=2.9.3
djangorestframework>=3.14.0
django-environ>=0.8.0
liquibase-core>=4.17.0